UI moved to properties modules
