+++ properties/object.py
@@ -2,6 +2,9 @@ from .base_classes import RendermanPropertyGroup
 from .rib_helpers import rib
 from bpy.props import *
 from mathutils import Matrix
+from ..ui.base_classes import PRManPanel
+from bpy.types import Panel
+from ..resources.icons.icons import load_icons
 
 ''' Object Properties '''
 
@@ -195,3 +198,85 @@ class RendermanObjectSettings(RendermanPropertyGroup):
         m = s * r * l
 
         ri.Transform(rib(m))
+
+
+class OBJECT_PT_renderman_object_geometry(PRManPanel, Panel):
+    bl_context = "object"
+    bl_label = "Renderman Geometry"
+
+    def draw(self, context):
+        layout = self.layout
+        ob = context.object
+        rm = ob.renderman
+        anim = rm.archive_anim_settings
+
+        col = layout.column()
+        col.prop(rm, "geometry_source")
+
+        if rm.geometry_source in ('ARCHIVE', 'DELAYED_LOAD_ARCHIVE'):
+            col.prop(rm, "path_archive")
+
+            col.prop(anim, "animated_sequence")
+            if anim.animated_sequence:
+                col.prop(anim, "blender_start")
+                row = col.row()
+                row.prop(anim, "sequence_in")
+                row.prop(anim, "sequence_out")
+
+        elif rm.geometry_source == 'PROCEDURAL_RUN_PROGRAM':
+            col.prop(rm, "path_runprogram")
+            col.prop(rm, "path_runprogram_args")
+        elif rm.geometry_source == 'DYNAMIC_LOAD_DSO':
+            col.prop(rm, "path_dso")
+            col.prop(rm, "path_dso_initial_data")
+
+        if rm.geometry_source in ('DELAYED_LOAD_ARCHIVE',
+                                  'PROCEDURAL_RUN_PROGRAM',
+                                  'DYNAMIC_LOAD_DSO'):
+            col.prop(rm, "procedural_bounds")
+
+            if rm.procedural_bounds == 'MANUAL':
+                colf = layout.column_flow()
+                colf.prop(rm, "procedural_bounds_min")
+                colf.prop(rm, "procedural_bounds_max")
+
+        if rm.geometry_source == 'BLENDER_SCENE_DATA':
+            col.prop(rm, "primitive")
+
+            colf = layout.column_flow()
+
+            if rm.primitive in ('CONE', 'DISK'):
+                colf.prop(rm, "primitive_height")
+            if rm.primitive in ('SPHERE', 'CYLINDER', 'CONE', 'DISK'):
+                colf.prop(rm, "primitive_radius")
+            if rm.primitive == 'TORUS':
+                colf.prop(rm, "primitive_majorradius")
+                colf.prop(rm, "primitive_minorradius")
+                colf.prop(rm, "primitive_phimin")
+                colf.prop(rm, "primitive_phimax")
+            if rm.primitive in ('SPHERE', 'CYLINDER', 'CONE', 'TORUS'):
+                colf.prop(rm, "primitive_sweepangle")
+            if rm.primitive in ('SPHERE', 'CYLINDER'):
+                colf.prop(rm, "primitive_zmin")
+                colf.prop(rm, "primitive_zmax")
+            if rm.primitive == 'POINTS':
+                colf.prop(rm, "primitive_point_type")
+                colf.prop(rm, "primitive_point_width")
+
+            # col.prop(rm, "export_archive")
+            # if rm.export_archive:
+            #    col.prop(rm, "export_archive_path")
+
+        rman_archive = load_icons().get("archive_RIB")
+        col = layout.column()
+        col.operator("export.export_rib_archive",
+                     text="Export Object as RIB Archive.", icon_value=rman_archive.icon_id)
+
+        col = layout.column()
+        # col.prop(rm, "export_coordsys")
+
+        row = col.row()
+        row.prop(rm, "motion_segments_override", text="")
+        sub = row.row()
+        sub.active = rm.motion_segments_override
+        sub.prop(rm, "motion_segments")
