--- ui/render_advanced_settings.py
@@ -1,91 +1,0 @@
-# ##### BEGIN MIT LICENSE BLOCK #####
-#
-# Copyright (c) 2015 - 2017 Pixar
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-#
-#
-# ##### END MIT LICENSE BLOCK #####
-
-import bpy
-from .base_classes import PRManPanel
-from bpy.types import Panel
-from ..resources.icons.icons import load_icons
-
-
-class RENDER_PT_renderman_advanced_settings(PRManPanel, Panel):
-    '''This panel covers additional render settings
-
-    # shading and tessellation
-    # geometry caches
-    # pixel filter
-    # render tiled order
-    # additional options (statistics, rib and texture generation caching,
-    thread settings)'''
-    bl_context = "render"
-    bl_label = "Advanced"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    def draw(self, context):
-        layout = self.layout
-        scene = context.scene
-        rm = scene.renderman
-
-        layout.separator()
-
-        col = layout.column()
-        col.label("Shading and Tessellation:")
-        col.prop(rm, "micropoly_length")
-        col.prop(rm, "dicing_strategy")
-        row = col.row()
-        row.enabled = rm.dicing_strategy == "worlddistance"
-        row.prop(rm, "worlddistancelength")
-        col.prop(rm, "instanceworlddistancelength")
-
-        layout.separator()
-
-        col = layout.column()
-        col.label("Cache Settings:")
-        col.prop(rm, "texture_cache_size")
-        col.prop(rm, "geo_cache_size")
-        col.prop(rm, "opacity_cache_size")
-        layout.separator()
-        col = layout.column()
-        col.label("Pixel Filter:")
-        col.prop(rm, "pixelfilter")
-        row = col.row(align=True)
-        row.prop(rm, "pixelfilter_x", text="Size X")
-        row.prop(rm, "pixelfilter_y", text="Size Y")
-        layout.separator()
-        col = layout.column()
-        col.label("Bucket Order:")
-        col.prop(rm, "bucket_shape")
-        if rm.bucket_shape == 'SPIRAL':
-            row = col.row(align=True)
-            row.prop(rm, "bucket_sprial_x", text="X")
-            row.prop(rm, "bucket_sprial_y", text="Y")
-        layout.separator()
-        col = layout.column()
-        row = col.row()
-        row.prop(rm, "use_statistics", text="Output stats")
-        col.operator('rman.open_rib')
-        row = col.row()
-        col.prop(rm, "always_generate_textures")
-        col.prop(rm, "lazy_rib_gen")
-        col.prop(rm, "threads")
